requirements.txt
bzip2=1.0.8=h2bbff1b_6
ca-certificates=2024.3.11=haa95532_0
libffi=3.4.4=hd77b12b_1
openssl=3.0.13=h2bbff1b_2
pip=22.3.1=py311haa95532_0
python=3.11.9=he1021f5_0
sqlite=3.45.3=h2bbff1b_0
tk=8.6.14=h0416ee5_0
vc=14.2=h2eaa2aa_1
vs2015_runtime=14.29.30133=h43f2093_3
xz=5.4.6=h8cc25b3_1
zlib=1.2.13=h8cc25b3_1
absl-py==2.1.0
accelerate==0.31.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
clarabel==0.9.0
click==8.1.7
cmeel==0.53.3
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cvxopt==1.3.2
cycler==0.12.1
daqp==0.5.1.post1
datasets==2.20.0
debugpy==1.8.1
decorator==5.1.1
deprecated==1.2.14
dill==0.3.8
distro==1.9.0
dynaconf==3.2.6
ecos==2.0.14
et-xmlfile==1.1.0
evaluate==0.4.2
evidently==0.4.33
executing==2.0.1
faker==26.0.0
filelock==3.15.3
flatbuffers==24.3.25
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.5.0
fuzzywuzzy==0.18.0
gast==0.5.4
gensim==4.3.3
google-pasta==0.2.0
googleapis-common-protos==1.63.2
grpcio==1.64.1
h11==0.14.0
h5py==3.11.0
highspy==1.7.1
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.7
importlib-metadata==8.0.0
intel-openmp==2021.4.0
ipykernel==6.29.4
ipython==8.25.0
iterative-telemetry==0.0.8
jedi==0.19.1
jinja2==3.1.4
joblib==1.4.2
jupyter-client==8.6.2
jupyter-core==5.7.2
keras==3.3.3
kiwisolver==1.4.5
libclang==18.1.1
litestar==2.10.0
lxml==5.2.2
markdown==3.6
markdown-it-py==3.0.0
markupsafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mkl==2021.4.0
ml-dtypes==0.3.2
mpmath==1.3.0
msgspec==0.18.6
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
numpy==1.26.4
openpyxl==3.1.2
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-exporter-otlp-proto-http==1.26.0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
opt-einsum==3.3.0
optree==0.11.0
osqp==0.6.7.post0
packaging==24.0
pandas==2.2.2
parso==0.8.4
patsy==0.5.6
pillow==10.3.0
piqp==0.4.1
platformdirs==4.2.2
plotly==5.23.0
polyfactory==2.16.2
prompt-toolkit==3.0.45
protobuf==4.25.3
proxsuite==0.6.6
psutil==5.9.8
pure-eval==0.2.2
pyarrow==16.1.0
pyarrow-hotfix==0.6
pydantic==2.8.2
pydantic-core==2.20.1
pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0
python-dotenv==1.0.1
python-http-client==3.3.7
pytz==2024.1
pywin32==306
pyyaml==6.0.1
pyzmq==26.0.3
qdldl==0.1.7.post4
qpalm==1.2.3
qpsolvers==4.3.2
quadprog==0.1.12
regex==2024.5.15
requests==2.32.3
rich==13.7.1
rich-click==1.8.3
safetensors==0.4.3
scikit-base==0.8.2
scikit-learn==1.4.2
scipy==1.13.0
scs==3.2.5
seaborn==0.13.2
sendgrid==6.11.0
setuptools==65.5.1
shellingham==1.5.4
six==1.16.0
sktime==0.31.0
smart-open==7.0.4
sniffio==1.3.1
stack-data==0.6.3
starkbank-ecdsa==2.2.0
statsmodels==0.14.2
sympy==1.12.1
tbb==2021.13.0
tenacity==9.0.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
tf-keras==2.16.0
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.3.1
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
transformers==4.41.2
typer==0.12.3
typing-extensions==4.12.0
typing-inspect==0.9.0
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.3
watchdog==4.0.1
watchfiles==0.22.0
wcwidth==0.2.13
websockets==12.0
werkzeug==3.0.3
wheel==0.38.4
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.19.2
